Split out BatteryMonitor
