+++ src/board/drivers/mod.rs
@@ -2,5 +2,6 @@
 pub mod battery_adc;
 #[cfg(feature = "battery_max17055")]
 pub mod battery_fg;
+pub mod battery_monitor;
 pub mod display;
 pub mod frontend;
